# api/openapi.yaml
openapi: 3.1.0
info:
  title: Rackd API
  description: Infrastructure management API for devices, networks, datacenters, and discovery
  version: 1.0.0
  contact:
    name: Rackd
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        maximum: 1000
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: >
            One of: INVALID_JSON, INVALID_INPUT, VALIDATION_ERROR, NOT_FOUND,
            FORBIDDEN, UNAUTHORIZED, INVALID_CREDENTIALS, ALREADY_EXISTS,
            IP_NOT_AVAILABLE, CANNOT_DELETE_SELF, SYSTEM_ROLE,
            RATE_LIMIT_EXCEEDED, LOGIN_RATE_LIMIT_EXCEEDED, CSRF_FAILED,
            EXPIRED_KEY, LEGACY_API_KEY_UNSUPPORTED, INTERNAL_ERROR
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Network:
      type: object
      required: [id, name, subnet, vlan_id, datacenter_id, description, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        subnet: { type: string, description: "CIDR notation" }
        vlan_id: { type: integer }
        datacenter_id: { type: string, format: uuid }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      additionalProperties: true
    NetworkInput:
      type: object
      required: [name, subnet]
      properties:
        name: { type: string, maxLength: 255 }
        subnet:
          type: string
          description: CIDR notation (e.g. 10.0.0.0/24); validated on create and update
        vlan_id: { type: integer, minimum: 0, maximum: 4094 }
        datacenter_id: { type: string, format: uuid }
        description: { type: string, maxLength: 4096 }
    NetworkPool:
      type: object
      required: [id, network_id, name, start_ip, end_ip, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        network_id: { type: string, format: uuid }
        name: { type: string }
        start_ip: { type: string }
        end_ip: { type: string }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      additionalProperties: true
    PoolInput:
      type: object
      required: [name, start_ip, end_ip]
      properties:
        name: { type: string }
        start_ip: { type: string }
        end_ip: { type: string }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
    NetworkUtilization:
      type: object
      required: [network_id, total_ips, used_ips, available_ips, utilization]
      properties:
        network_id: { type: string, format: uuid }
        total_ips: { type: integer }
        used_ips: { type: integer }
        available_ips: { type: integer }
        utilization: { type: number, format: float }
...
    IPHeatmap:
      type: object
      properties:
        pool_id: { type: string, format: uuid }
        entries:
          type: array
          items:
            type: object
            properties:
              ip: { type: string }
              status: { type: string, enum: [available, used, reserved] }
              device_id: { type: string, format: uuid }
      additionalProperties: true
...
    BulkResult:
      type: object
      properties:
        total: { type: integer }
        success: { type: integer }
        failed: { type: integer }
        errors:
          type: array
          items: { type: string }
    BulkDeleteRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/networks:
    get:
      operationId: listNetworks
      tags: [Networks]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: name
          in: query
          schema: { type: string }
        - name: datacenter_id
          in: query
          schema: { type: string, format: uuid }
        - name: vlan_id
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: List of networks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Network'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createNetwork
      tags: [Networks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/networks/{id}:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getNetwork
      tags: [Networks]
      responses:
        '200':
          description: Network details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updateNetwork
      tags: [Networks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteNetwork
      tags: [Networks]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/networks/{id}/utilization:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getNetworkUtilization
      tags: [Networks]
      responses:
        '200':
          description: Utilization stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkUtilization'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/networks/{id}/pools:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: listNetworkPools
      tags: [Pools]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: Pools in network
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NetworkPool'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createNetworkPool
      tags: [Pools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkPool'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── Pools ──
...
  /api/pools/{id}/heatmap:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getPoolHeatmap
      tags: [Pools]
      responses:
        '200':
          description: IP heatmap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPHeatmap'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/networks/bulk:
    post:
      operationId: bulkCreateNetworks
      tags: [Bulk Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/NetworkInput'
      responses:
        '200':
          description: Bulk result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResult'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: bulkDeleteNetworks
      tags: [Bulk Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteRequest'
      responses:
        '200':
          description: Bulk result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResult'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── Search ──
...